djreact/requirements.txt
Django==1.11.20
django-webpack-loader
djangorestframework==3.9.2
factory_boy==2.5.2
mysqlclient==1.3.7
virtualenv
pytest==4.4.1